+++ payment/models.py
@@ -5,7 +5,7 @@ from students.models import Enrollment
 class PaymentModel(models.Model):
     enrollment = models.ForeignKey(Enrollment, on_delete=models.CASCADE, verbose_name="Запись")
     lessons_covered = models.PositiveIntegerField(default=0, verbose_name="Кол-во покрытых уроков")
-    amount = models.DecimalField(max_digits=10, decimal_places=2, null=True, verbose_name="Сумма оплаты")
+    amount = models.FloatField(null=True, verbose_name="Сумма оплаты")
     date = models.DateTimeField(auto_now_add=True, verbose_name="Дата и время оплаты")
     status = models.BooleanField(default=False, verbose_name="Статус Оплаты")
